Release v.2.5 res - tructure
