--- lib/core/screens/home_screen/home_controller.dart
@@ -1,72 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:nautik_app/core/global/global.dart';
-
-void logoutUser(BuildContext context) {
-  showCupertinoDialog(
-    context: context,
-    builder: (x) {
-      return AlertDialog(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-        title: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Text(
-              textAlign: TextAlign.center,
-              'Cerrar sesión',
-              style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold),
-            ),
-          ],
-        ),
-        content: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text(
-                  'Esta seguro que desea cerrar \nsesion?',
-                  style: TextStyle(fontSize: generalText),
-                  textAlign: TextAlign.center,
-                ),
-              ],
-            ),
-          ],
-        ),
-        actions: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Row(
-                children: [
-                  TextButton(
-                    onPressed: () {
-                      Navigator.pop(context);
-                    },
-                    child: Text(
-                      'Cancelar',
-                      style: TextStyle(fontSize: generalText),
-                    ),
-                  ),
-                  SizedBox(width: 20),
-                  TextButton(
-                    onPressed: () {
-                      Navigator.pushReplacementNamed(context, '/Login');
-                    },
-                    child: Text(
-                      'Aceptar',
-                      style: TextStyle(
-                        color: primaryColor,
-                        fontSize: generalText,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ],
-      );
-    },
-  );
-}
