+++ lib/core/utils/ui_helpers.dart
@@ -7,10 +7,3 @@ Widget buildHeight(double size) {
 Widget buildWidth(double size) {
   return SizedBox(width: size);
 }
-
-bool onOff = false;
-int triedAuth = 0;
-
-var emailController = TextEditingController();
-var passwordController = TextEditingController();
-
