+++ lib/core/widgets/global_drawer.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
-import 'package:nautik_app/core/screens/home_screen/home_controller.dart';
-import 'package:nautik_app/core/screens/login_screen/login_model.dart';
+import 'package:nautik_app/core/utils/ui_helpers.dart';
+import 'package:nautik_app/widgets/custom_widgets.dart';
 
 class GlobalDrawer extends StatelessWidget {
   const GlobalDrawer({super.key});
