+++ lib/modules/auth/controllers/register_controller.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
 import 'package:nautik_app/core/global/global.dart';
-import 'package:nautik_app/core/screens/registro_screen/registro_model.dart';
+import 'package:nautik_app/modules/auth/models/user_model.dart';
+import 'package:nautik_app/modules/auth/widgets/form_widget.dart';
+import 'package:nautik_app/widgets/dialogs.dart';
 
 void registerUser(BuildContext context) {
   if (emailControllerVR.text.isEmpty ||
