+++ lib/routes/routes.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:nautik_app/core/screens/categorias_screen/categorias_screen.dart';
-import 'package:nautik_app/core/screens/home_screen/home_screen.dart';
-import 'package:nautik_app/core/screens/login_screen/login_screen.dart';
-import 'package:nautik_app/core/screens/registro_screen/registro_screen.dart';
+import 'package:nautik_app/screens/categorias_screen.dart';
+import 'package:nautik_app/screens/home_screen.dart';
+import 'package:nautik_app/screens/login_screen.dart';
+import 'package:nautik_app/screens/registro_screen.dart';
 
 Map<String, Widget Function(BuildContext)> routes = {
   '/Login': (context) => Login(),
