+++ lib/screens/carrito_screen.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
 import 'package:nautik_app/core/global/global.dart';
-import 'package:nautik_app/core/screens/login_screen/login_model.dart';
-import 'package:nautik_app/core/screens/mapa_screen/mapa_screen.dart';
+import 'package:nautik_app/core/utils/ui_helpers.dart';
+import 'package:nautik_app/screens/mapa_screen.dart';
+import 'package:nautik_app/themes/colors.dart';
+import 'package:nautik_app/widgets/custom_widgets.dart';
 
 class Carritoview extends StatelessWidget {
   final List<Map<String, dynamic>> productos;
