+++ lib/screens/categorias_screen.dart
@@ -1,10 +1,9 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
-import 'package:nautik_app/core/global/global.dart';
-import 'package:nautik_app/core/screens/login_screen/login_model.dart';
-import 'package:nautik_app/core/screens/productos_screen/productos_screen.dart';
+import 'package:nautik_app/core/utils/ui_helpers.dart';
+import 'package:nautik_app/modules/products/data/mock_menu_data.dart';
+import 'package:nautik_app/screens/productos_screen.dart';
+import 'package:nautik_app/themes/colors.dart';
 
 class Categoriesview extends StatefulWidget {
   const Categoriesview({super.key});
