+++ lib/themes/colors.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+Color primaryColor = const Color.fromARGB(255, 19, 154, 248);
+Color backgroundColor = Colors.white;
+double generalText = 17;
+Color searchbarColor = Color.fromARGB(255, 240, 240, 240);
