+++ lib/widgets/custom_widgets.dart
@@ -1,11 +1,11 @@
 // ignore_for_file: file_names
 
-import 'dart:async';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
-import 'package:nautik_app/core/global/global.dart';
-import 'package:nautik_app/core/screens/login_screen/login_model.dart';
+import 'package:nautik_app/core/utils/ui_helpers.dart';
+import 'package:nautik_app/modules/products/data/mock_menu_data.dart';
+import 'package:nautik_app/themes/colors.dart';
 
 void welcomeUser(BuildContext context) {
   showDialog(
@@ -140,5 +140,91 @@ void buildDialogOnline(BuildContext context) {
         ],
       );
     },
+  );
+}
+
+void logoutUser(BuildContext context) {
+  showCupertinoDialog(
+    context: context,
+    builder: (x) {
+      return AlertDialog(
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+        title: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              textAlign: TextAlign.center,
+              'Cerrar sesión',
+              style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold),
+            ),
+          ],
+        ),
+        content: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Text(
+                  'Esta seguro que desea cerrar \nsesion?',
+                  style: TextStyle(fontSize: generalText),
+                  textAlign: TextAlign.center,
+                ),
+              ],
+            ),
+          ],
+        ),
+        actions: [
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Row(
+                children: [
+                  TextButton(
+                    onPressed: () {
+                      Navigator.pop(context);
+                    },
+                    child: Text(
+                      'Cancelar',
+                      style: TextStyle(fontSize: generalText),
+                    ),
+                  ),
+                  SizedBox(width: 20),
+                  TextButton(
+                    onPressed: () {
+                      Navigator.pushReplacementNamed(context, '/Login');
+                    },
+                    child: Text(
+                      'Aceptar',
+                      style: TextStyle(
+                        color: primaryColor,
+                        fontSize: generalText,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ],
+      );
+    },
+  );
+}
+
+Widget buildPrimaryButton(BuildContext context, String text, Function onTap) {
+  return ElevatedButton(
+    style: ElevatedButton.styleFrom(
+      fixedSize: Size(300, 55),
+      backgroundColor: primaryColor,
+      foregroundColor: Colors.white,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+    ),
+    onPressed: () => onTap(),
+    child: Text(
+      text,
+      style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+    ),
+    
   );
 }
