+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   shared_preferences: ^2.5.3
   flutter_map: ^8.1.1
   latlong2: ^0.9.1
+  provider: ^6.1.4
 
 dev_dependencies:
   flutter_test:
